refactor: centralise display filter logic to reduce duplicated fn calls
